ref(core): Move SQL sanitization to `utils/sql`
`_sanitizeSqlQuery` lived in the postgres.js integration, but it is dialect-generic and now
serves the mysql, mysql2, postgres, postgres.js and Cloudflare D1 paths. Move it next to
`getSqlQuerySummary`, which is its only downstream consumer and already lives there, so the
sanitize-then-summarize pipeline reads as one unit and D1 no longer reaches into a postgres
integration module for it.

Rename `_sanitizeSqlQuery` to `sanitizeSqlQuery` to match its neighbour; the public
`_INTERNAL_sanitizeSqlQuery` alias is unchanged, so nothing outside core moves. Tests move
with it, and the leak regressions can now assert the summary directly instead of only the
sanitized statement. No behavior change.
