+++ packages/core/src/server-exports.ts
@@ -29,14 +29,17 @@ export type {
 /* oxlint-enable typescript/no-deprecated */
 export {
   instrumentPostgresJsSql,
-  _sanitizeSqlQuery as _INTERNAL_sanitizeSqlQuery,
   _reconstructQuery as _INTERNAL_reconstructPostgresQuery,
   _buildConnectionContext as _INTERNAL_buildPostgresConnectionContext,
   _getConnectionAttributes as _INTERNAL_getConnectionAttributes,
   _getOperationName as _INTERNAL_getPostgresOperationName,
 } from './integrations/postgresjs';
 export type { PostgresConnectionContext } from './integrations/postgresjs';
-export { getSqlQuerySummary as _INTERNAL_getSqlQuerySummary } from './utils/sql';
+export {
+  getSqlQuerySummary as _INTERNAL_getSqlQuerySummary,
+  sanitizeSqlQuery as _INTERNAL_sanitizeSqlQuery,
+} from './utils/sql';
+export type { SqlDialect } from './utils/sql';
 
 export { patchHttpModuleClient } from './integrations/http/client-patch';
 export { getHttpClientSubscriptions } from './integrations/http/client-subscriptions';
